+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- \[mpm\] Fix import from private `click.extra` submodule.
+
 ## {gh}`5.6.0 (2022-09-26) <compare/v5.5.1...v5.6.0>`
 
 - \[brew,cask\] Add support for `remove` operation in homebrew.